dws维度退化5张表完成

+++ question/src/main/scala/com/mxl/education/question/bean/Models.scala
@@ -413,7 +413,7 @@ case class DwsQzPaper(
 	                     paperyear: String,
 	                     suitnum: String,
 	                     papername: String,
-	                     totalscore: BigDecimal,
+	                     totalscore: Float,
 	                     chapterid: Int,
 	                     chapterlistid: Int,
 	                     dt: String,
